pyproject.toml
[project]
name = "urlshortener"
version = "0.1.0"
description = ""
authors = [
    {name = "Oleksandr Burdeinyi",email = "bonus05@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "drf-spectacular (>=0.27.2,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
